pyproject.toml
[tool.poetry]
name = "bingo_reader"
version = "0.1.0"
description = ""
authors = ["Christian Jørgensen <friis93@me.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pytesseract = "^0.3.2"
Pillow = "^7.0.0"
autopep8 = "^1.5"
jupyter = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
